Cleaned up a few functions, fixed logging, etc.
